+++ __test__/buildData.test.ts
@@ -2,19 +2,30 @@ import { CVCSVCLI } from "../src/cvcsvcli";
 import * as readline from "readline";
 describe("buildData", () => {
   let subject: CVCSVCLI;
-  let mockStdoutClearLine: jest.SpyInstance<
-    boolean,
-    [NodeJS.WritableStream, readline.Direction, (() => void)?]
-  >;
+  let mockStdoutClearLine: { (): typeof jest; mockRestore?: any };
   let mockStdoutWrite: jest.SpyInstance<
     boolean,
-    [string | Uint8Array, string?, ((err?: Error) => void)?]
+    [
+      string | Uint8Array,
+      (
+        | "ascii"
+        | "utf8"
+        | "utf-8"
+        | "utf16le"
+        | "ucs2"
+        | "ucs-2"
+        | "base64"
+        | "latin1"
+        | "binary"
+        | "hex"
+        | undefined
+      )?,
+      (((err?: Error | undefined) => void) | undefined)?
+    ]
   >;
 
   beforeAll(() => {
-    mockStdoutClearLine = jest
-      .spyOn(readline, "clearLine")
-      .mockImplementation(() => true);
+    mockStdoutClearLine = jest.mock("readline").autoMockOn;
     mockStdoutWrite = jest
       .spyOn(process.stdout, "write")
       .mockImplementation(() => true);
@@ -23,7 +34,7 @@ describe("buildData", () => {
       bucketName: "foo",
       rootDirectory: "foo",
       productCategory: "packagedgoods-v1",
-      productSet: "test"
+      productSet: "test",
     };
   });
   afterAll(() => {
@@ -38,7 +49,7 @@ describe("buildData", () => {
         "product-id": "TESTING",
         "product-display-name": "Testing",
         "product-category": "packagedgoods-v1",
-        "product-set-id": "test"
+        "product-set-id": "test",
       },
       {
         "image-uri": "gs://foo/images/another_dir/2.png",
@@ -46,8 +57,8 @@ describe("buildData", () => {
         "product-id": "ANOTHERDIR",
         "product-display-name": "Another Dir",
         "product-category": "packagedgoods-v1",
-        "product-set-id": "test"
-      }
+        "product-set-id": "test",
+      },
     ];
     const result = await subject["buildData"]();
     expect(result).toEqual(expected);